Renames associated activity component
