+++ src/app/modules/group/components/associated-item/associated-item.component.scss
