+++ src/app/modules/group/pages/group-edit/group-edit.component.html
@@ -30,8 +30,8 @@
           </div>
         </div>
       </alg-section>
-      <alg-associated-activity formControlName="rootActivity" contentType="activity"></alg-associated-activity>
-      <alg-associated-activity formControlName="rootSkill" contentType="skill"></alg-associated-activity>
+      <alg-associated-item formControlName="rootActivity" contentType="activity"></alg-associated-item>
+      <alg-associated-item formControlName="rootSkill" contentType="skill"></alg-associated-item>
 
       <div class="remove-button-section">
         <alg-group-remove-button [group]="state.data" (groupDeleted)="refreshNav()"></alg-group-remove-button>
